Extract session overlay viewport logic into dedicated module
Move viewport-drift detection, device type classification, text-editing
checks, and top-chrome inset computation from useVisualViewport.js into
sessionOverlayViewport.js for better separation of concerns. The original
module now re-exports from the new one to preserve the public API.
