+++ packages/web/src/composables/useVisualViewport.js
@@ -1,7 +1,20 @@
 import { onMounted, onUnmounted } from 'vue';
 import { computeSessionOverlayKeyboardBottomInset } from './sessionOverlayKeyboardInset.js';
+import {
+  checkOverlayViewportDrift as checkOverlayViewportDriftElement,
+  clearOverlayViewportDrift,
+  computeSessionOverlayTopChromeInset,
+  isActiveTextEditing,
+  isTextEditingElement,
+} from './sessionOverlayViewport.js';
 
 export { computeSessionOverlayKeyboardBottomInset } from './sessionOverlayKeyboardInset.js';
+export {
+  clearOverlayViewportDrift,
+  computeSessionOverlayTopChromeInset,
+  isActiveTextEditing,
+  isTextEditingElement,
+} from './sessionOverlayViewport.js';
 
 let rafId = null;
 let settleRafId = null;
@@ -11,11 +24,6 @@ let settleLastRect = null;
 let settleStableSamples = 0;
 let isSessionOverlayPromptFocused = false;
 
-const SESSION_OVERLAY_TOP_CHROME_THRESHOLD = 64;
-const KEYBOARD_HEIGHT_DELTA_THRESHOLD = 120;
-const KEYBOARD_VIEWPORT_RATIO_THRESHOLD = 0.85;
-const TABLET_MIN_LAYOUT_DIMENSION = 700;
-
 function getPixelValue(value, fallback) {
   return Number.isFinite(value) ? `${value}px` : fallback;
 }
@@ -23,120 +31,6 @@ function getPixelValue(value, fallback) {
 function getVisualViewportRect() {
   const { offsetTop, height } = window.visualViewport;
   return { offsetTop, height };
-}
-
-function isValidOffsetTop(offsetTop) {
-  return (
-    Number.isFinite(offsetTop) &&
-    offsetTop > 0 &&
-    offsetTop <= SESSION_OVERLAY_TOP_CHROME_THRESHOLD
-  );
-}
-
-function getDeviceType(userAgent, platform, maxTouchPoints) {
-  const ua = String(userAgent);
-  const devicePlatform = String(platform);
-  const touchPoints = Number(maxTouchPoints) || 0;
-  const isAndroid = /Android/i.test(ua);
-  const isAndroidMobile = isAndroid && /Mobile/i.test(ua);
-  const isIPhoneLike = /iPhone|iPod/i.test(`${ua} ${devicePlatform}`);
-  const isIPad =
-    /iPad/i.test(`${ua} ${devicePlatform}`) ||
-    (devicePlatform === 'MacIntel' && touchPoints > 1);
-  const isAndroidTablet = isAndroid && !/Mobile/i.test(ua);
-
-  return {
-    isPhone: isIPhoneLike || isAndroidMobile,
-    isTablet: isIPad || isAndroidTablet,
-  };
-}
-
-function hasKeyboardShapedViewport(layoutHeight, visualViewportHeight) {
-  return (
-    Number.isFinite(layoutHeight) &&
-    layoutHeight > 0 &&
-    Number.isFinite(visualViewportHeight) &&
-    (layoutHeight - visualViewportHeight > KEYBOARD_HEIGHT_DELTA_THRESHOLD ||
-      visualViewportHeight / layoutHeight < KEYBOARD_VIEWPORT_RATIO_THRESHOLD)
-  );
-}
-
-function hasTabletSizedLayout(layoutWidth, layoutHeight) {
-  return (
-    Number.isFinite(layoutWidth) &&
-    Number.isFinite(layoutHeight) &&
-    Math.min(layoutWidth, layoutHeight) >= TABLET_MIN_LAYOUT_DIMENSION
-  );
-}
-
-const TEXT_LIKE_INPUT_TYPES = new Set([
-  '',
-  'text',
-  'search',
-  'email',
-  'url',
-  'tel',
-  'password',
-  'number',
-]);
-
-export function isTextEditingElement(element) {
-  if (!element || element.nodeType !== 1) {
-    return false;
-  }
-
-  const editableElement = element.closest?.('[contenteditable]');
-  if (editableElement) {
-    const contentEditable = editableElement.getAttribute('contenteditable');
-    return contentEditable === null || contentEditable.toLowerCase() !== 'false';
-  }
-
-  const tagName = element.tagName?.toLowerCase();
-  if (tagName === 'textarea') {
-    return true;
-  }
-
-  if (tagName !== 'input') {
-    return false;
-  }
-
-  return TEXT_LIKE_INPUT_TYPES.has((element.getAttribute('type') || '').toLowerCase());
-}
-
-export function isActiveTextEditing() {
-  if (typeof document === 'undefined') {
-    return false;
-  }
-  return isTextEditingElement(document.activeElement);
-}
-
-export function computeSessionOverlayTopChromeInset({
-  offsetTop,
-  visualViewportHeight,
-  layoutWidth,
-  layoutHeight,
-  userAgent = '',
-  platform = '',
-  maxTouchPoints = 0,
-}) {
-  if (!isValidOffsetTop(offsetTop)) {
-    return 0;
-  }
-
-  const deviceType = getDeviceType(userAgent, platform, maxTouchPoints);
-  if (deviceType.isPhone) {
-    return 0;
-  }
-
-  if (hasKeyboardShapedViewport(layoutHeight, visualViewportHeight)) {
-    return 0;
-  }
-
-  if (deviceType.isTablet || hasTabletSizedLayout(layoutWidth, layoutHeight)) {
-    return offsetTop;
-  }
-
-  return 0;
 }
 
 export function setSessionOverlayPromptFocus(focused) {
@@ -294,98 +188,8 @@ export function requestVisualViewportSettle(options = {}) {
   settleRafId = requestAnimationFrame(sample);
 }
 
-// ---------------------------------------------------------------------------
-// Overlay viewport-drift correction
-//
-// On iPad Safari the visual viewport can shift relative to the layout viewport
-// when the browser chrome (URL bar / tab bar) collapses or expands, during
-// scroll-bounce, or after tab switches. `position: fixed` elements follow the
-// *layout* viewport, so they drift off-screen even though JS APIs like
-// `getBoundingClientRect` and `window.scrollY` still report 0.
-//
-// `checkOverlayViewportDrift` reads `visualViewport.offsetTop` and pins the
-// overlay element to the visual viewport via inline styles when drift > threshold.
-// ---------------------------------------------------------------------------
-const DRIFT_THRESHOLD_PX = 2;
-
-/**
- * Check for visual-viewport drift and correct the given element's position.
- *
- * Intended to be called periodically (setInterval) and/or on visualViewport
- * scroll/resize events while a fullscreen overlay is open.
- *
- * The correction only applies on tablets / large-screen devices when the
- * software keyboard is NOT open. On phones the browser-chrome drift issue
- * doesn't occur, and when the keyboard is open the viewport offset is
- * expected (not drift) — repositioning the overlay would fight the
- * keyboard and push the focused input out of view.
- *
- * @param {HTMLElement|null} element  The overlay backdrop element to pin.
- */
 export function checkOverlayViewportDrift(element) {
-  if (!element) return;
-
-  if (isActiveTextEditing()) {
-    writeVisualViewportVariables();
-    return;
-  }
-
-  // Force window scroll to 0 — page should never scroll while overlay is open.
-  if (window.scrollY !== 0) {
-    window.scrollTo(0, 0);
-  }
-
-  if (!window.visualViewport) {
-    writeVisualViewportVariables();
-    return;
-  }
-
-  const { offsetTop, height } = window.visualViewport;
-  const layoutWidth = window.innerWidth;
-  const layoutHeight = window.innerHeight;
-
-  // Use the same guard logic as computeSessionOverlayTopChromeInset:
-  // skip phones entirely, skip when the keyboard is open, and only
-  // correct on tablets / large-screen devices.
-  const deviceType = getDeviceType(
-    window.navigator?.userAgent,
-    window.navigator?.platform,
-    window.navigator?.maxTouchPoints,
-  );
-
-  const shouldCorrect =
-    !deviceType.isPhone &&
-    !hasKeyboardShapedViewport(layoutHeight, height) &&
-    (deviceType.isTablet || hasTabletSizedLayout(layoutWidth, layoutHeight)) &&
-    offsetTop > DRIFT_THRESHOLD_PX;
-
-  if (shouldCorrect) {
-    // Drift detected — override CSS `inset: 0` with explicit position.
-    // Inline styles beat scoped-CSS specificity, so this wins over `inset`.
-    element.style.top = `${offsetTop}px`;
-    element.style.bottom = 'auto';
-    element.style.height = `${height}px`;
-  } else {
-    // No drift (or phone / keyboard open) — clear inline overrides so
-    // the CSS `inset: 0` rule applies normally.
-    element.style.top = '';
-    element.style.bottom = '';
-    element.style.height = '';
-  }
-
-  writeVisualViewportVariables();
-}
-
-/**
- * Clear any inline position overrides applied by `checkOverlayViewportDrift`.
- *
- * @param {HTMLElement|null} element  The overlay backdrop element.
- */
-export function clearOverlayViewportDrift(element) {
-  if (!element) return;
-  element.style.top = '';
-  element.style.bottom = '';
-  element.style.height = '';
+  checkOverlayViewportDriftElement(element, { writeVisualViewportVariables });
 }
 
 /**
